ensure float in unit conversion

+++ src/unit_conversion/to_pu.jl
@@ -1,13 +1,13 @@
 function _si_to_pu!(data::Dict{String,Any},
     params::Dict{Symbol,Any}, nominal_values::Dict{Symbol,Any})
 
-    rescale_mass_flow = x -> x/nominal_values[:mass_flow]
-    rescale_mass_flux = x -> x/nominal_values[:mass_flux]
-    rescale_pressure = x -> x/nominal_values[:pressure]
-    rescale_length = x -> x/nominal_values[:length]
-    rescale_density = x -> x/nominal_values[:density]
-    rescale_diameter = x -> x/nominal_values[:length]
-    rescale_area = x -> x/nominal_values[:area]
+    rescale_mass_flow = x -> Float64(x/nominal_values[:mass_flow])
+    rescale_mass_flux = x -> Float64(x/nominal_values[:mass_flux])
+    rescale_pressure = x -> Float64(x/nominal_values[:pressure])
+    rescale_length = x -> Float64(x/nominal_values[:length])
+    rescale_density = x -> Float64(x/nominal_values[:density])
+    rescale_diameter = x -> Float64(x/nominal_values[:length])
+    rescale_area = x -> Float64(x/nominal_values[:area])
 
 
     function rescale_compressor_boundary_conditions!(value)
